refactor shared component structure
