+++ app/components/codeBlock/codeBlock.module.css
@@ -1,0 +1,3 @@
+.alert {
+  white-space: break-spaces;
+}
