+++ app/components/index.ts
@@ -1,5 +1,6 @@
 export { default as Button } from "./button/button";
 export { default as Card, CardPaper } from "./card/card";
+export { default as CodeBlock } from "./codeBlock/codeBlock";
 export { default as Layout } from "./layout/layout";
 export { default as Notification } from "./notification/notification";
 export { default as Reader } from "./reader/reader";
