+++ app/utils/index.ts
@@ -1,3 +1,4 @@
+export * from "./components";
 export * from "./format";
 export * from "./frontend";
 export * from "./hooks";
